pyproject.toml
[tool.poetry]
name = "airbyte-dbt-analytics"
version = "0.1.0"
description = ""
authors = ["Artur Tiburski Vaz Costa <artur_vazcosta@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.11.5"
apache-airflow = "^2.9.1"
apache-airflow-providers-dbt-cloud = "^3.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
